Refactoring, again

+++ nRFMeshProvision/Classes/Mesh Model/KeyIndex.swift
@@ -7,11 +7,11 @@
 
 import Foundation
 
-public typealias KeyIndex = Int
+public typealias KeyIndex = UInt16
 
 public extension KeyIndex {
     
-    /// A Key Index is 24-bit long Unsigned Integer.
+    /// A Key Index is 12-bit long Unsigned Integer.
     /// This property returns `true` if the value is in range 0...4095.
     var isValidKeyIndex: Bool {
         return self >= 0 && self <= 4095
